ffmpegServer: updated to build against latest release of areaDetector 1-9.

+++ ffmpegServerApp/src/ffmpegServer.h
@@ -74,7 +74,7 @@ private:
 
     NDArray* get_jpeg();
     NDArray* wait_for_jpeg(int sid);    
-    void allocScArray(int size);
+    void allocScArray(size_t size);
 };
 #define NUM_FFMPEG_SERVER_PARAMS (&LAST_FFMPEG_SERVER_PARAM - &FIRST_FFMPEG_SERVER_PARAM + 1)
